# Conflicts:
#	app/Http/Controllers/V2/InstanceController.php
#	app/Http/Requests/V2/CreateInstanceRequest.php
#	app/Http/Requests/V2/UpdateInstanceRequest.php
#	app/Models/V2/Instance.php
#	app/Resources/V2/InstanceResource.php
#	bootstrap/app.php

+++ app/Http/Controllers/V2/InstanceController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\V2\CreateInstanceRequest;
 use App\Http\Requests\V2\UpdateInstanceRequest;
 use App\Models\V2\Instance;
 use App\Resources\V2\InstanceResource;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use UKFast\DB\Ditto\QueryTransformer;
 
@@ -54,6 +55,9 @@ class InstanceController extends BaseController
     {
         $instance = new Instance();
         $instance->fill($request->only($instance->getFillableMinusKey()));
+        if (!$request->has('locked')) {
+            $instance->locked = false;
+        }
         $instance->save();
         $instance->refresh();
         return $this->responseIdMeta($request, $instance->getKey(), 201);
@@ -67,6 +71,11 @@ class InstanceController extends BaseController
     public function update(UpdateInstanceRequest $request, string $instanceId)
     {
         $instance = Instance::forUser(app('request')->user)->findOrFail($instanceId);
+        if (!$this->isAdmin &&
+            (!$request->has('locked') || $request->get('locked') !== false) &&
+            $instance->locked === true) {
+            return $this->isLocked();
+        }
         $instance->fill($request->only($instance->getFillableMinusKey()));
         $instance->save();
         return $this->responseIdMeta($request, $instance->getKey(), 200);
@@ -80,7 +89,24 @@ class InstanceController extends BaseController
     public function destroy(Request $request, string $instanceId)
     {
         $instance = Instance::forUser($request->user)->findOrFail($instanceId);
+        if (!$this->isAdmin && $instance->locked === true) {
+            return $this->isLocked();
+        }
         $instance->delete();
         return response()->json([], 204);
+    }
+
+    /**
+     * @return \Illuminate\Http\JsonResponse
+     */
+    private function isLocked(): JsonResponse
+    {
+        return JsonResponse::create([
+            'errors' => [
+                'title'  => 'Forbidden',
+                'detail' => 'The specified instance is locked',
+                'status' => 403,
+            ]
+        ], 403);
     }
 }
